feat: 键盘删除确认

+++ packages/js/mixins/multipleSelectMixin.js
@@ -26,30 +26,44 @@ export default {
       deleteItem: 'delItem'
     }),
     keydown (event) {
-      // 关闭默认事件
-      event.preventDefault()
       if (event.keyCode === 37) {
+        // 关闭默认事件
+        event.preventDefault()
         // 左箭头键被按下
         this.changeActivePos({ diffX: -1, diffY: 0 })
       } else if (event.keyCode === 38) {
+        // 关闭默认事件
+        event.preventDefault()
         // 上箭头键被按下
         this.changeActivePos({ diffX: 0, diffY: -1 })
       } else if (event.keyCode === 39) {
+        // 关闭默认事件
+        event.preventDefault()
         // 右箭头键被按下
         this.changeActivePos({ diffX: 1, diffY: 0 })
       } else if (event.keyCode === 40) {
+        // 关闭默认事件
+        event.preventDefault()
         // 下箭头键被按下
         this.changeActivePos({ diffX: 0, diffY: 1 })
       }
       // 删除
       if (event.keyCode === 8 || event.keyCode === 46) {
-        // 批量删除
-        if (Array.isArray(this.activeCodes) && this.activeCodes.length > 0) {
-          this.deleteItem(this.activeCodes)
-        } else {
-          // 单个删除
-          this.deleteItem(this.activeChart)
-        }
+        // 关闭默认事件
+        event.preventDefault()
+        this.$confirm('确定删除该组件吗？', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          // 批量删除
+          if (Array.isArray(this.activeCodes) && this.activeCodes.length > 0) {
+            this.deleteItem(this.activeCodes)
+          } else {
+            // 单个删除
+            this.deleteItem(this.activeChart)
+          }
+        }).catch(() => {})
       }
 
       if (event.shiftKey) {
